Update buttons with links
