+++ components/Tactics/TacticsCard/index.js
@@ -1,7 +1,5 @@
-import {
-  StyledArticle,
-  StyledDetailsLink,
-} from "../../Competition/CompetitionCard/StyledCompetitionCard";
+import { StyledArticle } from "../../Competition/CompetitionCard/StyledCompetitionCard";
+import { StyledLinkComponent } from "../../GeneralComponents/Links";
 
 export default function TacticCard({ tactic, path }) {
   if (!tactic) {
@@ -12,9 +10,9 @@ export default function TacticCard({ tactic, path }) {
     <StyledArticle>
       <h3>{tactic.name}</h3>
       <h4>Formation: {tactic.formation}</h4>
-      <StyledDetailsLink href={`/tactics/${tactic.id}/?${path}`}>
+      <StyledLinkComponent type="view" href={`/tactics/${tactic.id}/?${path}`}>
         View Details
-      </StyledDetailsLink>
+      </StyledLinkComponent>
     </StyledArticle>
   );
 }
